Fix exploration script and model.py implementation
Set initialization to 1 for gamma and beta.
Modify exploration script 'constmax_training' to actually utilize
ConSMax (before was running softmax).

The new curve is different, and very interesting showing adjustment of
the hyperparameters.

The performance is still better than softermax.

It was inconclusive whether the performance was at par or better than
regular softmax, we may require a larger dataset, as overfitting of this
dataset started right around where ConSMax overtook softermax...