Add fire positional encoding
This is just a draft commit

Because the repo is moving so quickly would be easier to apply the
changes on top of the latest.

Also including stability improvements for working with fire for
constantmax and polymax. Was thinking we might also want to try adding
normalization between the fire and softmax stages to test if this fixes
the compatibility.