Add feature to set random weights for each dataset
One of the options covered by this is to do just one dataset, while
leaving the others as validation only.

Next step is to associate the index of the chosen dataset with an
additional option to use this index to set a one hot vector within the
model.py via an update_model method.