Add option to explore irregular transformer sizes
This is just a draft. Suspecting that adding a double width transformer
to the middle of the model will improve its NLP modeling capability

To test, we'll loop through adding a single layer to the 6 layer model,
and log which produces the best validation loss.

The residual stream has to be modified for this to work, will try a few
variations including:
1. skipping entirely
2. linear for upscaling/downscaling
3. combination of above.