Add prototype for dual context training
Including a labelled shakespeare_char_punct which has code for
converting labels per character (in this case vowel, punctuation,
consonant or other, but can also be POS).

Currently we default to averaging the losses for each lm_head.

Also this only has two streams and no method for selecting the head to
inference from.

This allows for concurrent lables of the