Add learned rotation MLP variation and skip res
Learned Rotation is a direct rotation learned to test hypothesis if we
can gain greater parameter efficiency by more directly learning
rotations in place of the MLP.

Also adding a script for searching relevant settings such as skip
residual options, as we may want to only add to the residual before or after
rotations.

Another feature we should add are separate residual streams for
attention outputs and Learned Rotations or MLP.

We haven't implemented sequence of MLP yet, might be interesting to add.