Add learned rotations and sequential rotations
This is based on the possibility that we're learning a rotation to get
into the right dimension, then a separate rotation to get to the right
category.

So far generally doesn't do better than simple identity operation,
however thinking that we should do some hyperparameter sweeps on:

initial magnitudes (to get the tiling effect).
number of seqential rotations
number of heads

The idea is that we may need another rotation per head.